+++ CHANGES.rst
@@ -5,6 +5,10 @@ Changes for croud
 Unreleased
 ==========
 
+- Added the commands ``croud monitoring grafana create`` and
+  ``croud monitoring grafana remove`` These commands are used to create
+  and remove a Grafana dashboard for a project.
+
 0.10.0 - 2019/04/01
 ===================